requirements.txt
chardet==2.3.0
diff-match-patch==20121119
Django==2.1.3
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-formtools==2.1
django-import-export==1.1.0
django-mdeditor==0.1.9
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.2
django-simple-captcha==0.5.9
djangorestframework==3.9.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4
#language-selector==0.1
Markdown==3.0.1
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
Pygments==2.2.0
pygobject==3.20.0
python-apt==1.1.0b1
pytz==2018.7
PyYAML==3.13
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
tablib==0.12.1
ufw==0.35
unicodecsv==0.14.1
urllib3==1.13.1
uWSGI==2.0.17.1
xlrd==1.1.0
xlwt==1.3.0
